+++ src/esm/components/action/button/helpers/modifier-styles.js
@@ -1,137 +1,150 @@
 import { css } from 'styled-components';
-import COLORS from '../../../../constants/colors-config';
 export const modifierStyles = modifier => {
-  let borderColor;
-  let borderStyle = 'solid';
-  let borderSize = 'var(--border-width)';
-  let backgroundColor;
-  let color;
-  let hoverBorderColor = COLORS.primary2;
-  let hoverBgColor = COLORS.primary2;
-  let hoverColor = COLORS.background1;
-  let activeBorderColor = COLORS.primary3;
-  let activeBgColor = COLORS.primary3;
-  let activeColor = COLORS.background1;
-  let disabledBorderColor = 'var(--color-grey-200)';
-  let disabledBgColor = 'var(--color-grey-200)';
-  let disabledColor = 'var(--color-grey-600)';
+  const colors = {
+    borderColor: 'var(--color-grey-700)',
+    backgroundColor: 'var(--color-grey-000)',
+    color: 'var(--color-grey-900)',
+    hoverBorderColor: 'var(--color-primary-700)',
+    hoverBgColor: 'var(--color-primary-700)',
+    hoverColor: 'var(--color-grey-000)',
+    activeBorderColor: 'var(--color-primary-900)',
+    activeBgColor: 'var(--color-primary-900)',
+    activeColor: 'var(--color-grey-000)',
+    disabledBorderColor: 'var(--color-grey-200)',
+    disabledBgColor: 'var(--color-grey-200)',
+    disabledColor: 'var(--color-grey-600)'
+  };
 
   switch (modifier) {
     case 'hydrogen':
-      borderColor = 'var(--color-grey-700)';
-      backgroundColor = COLORS.background1;
-      color = COLORS.font1;
-      hoverBorderColor = 'var(--color-primary-500)';
-      hoverBgColor = COLORS.background1;
-      hoverColor = COLORS.font1;
-      activeBorderColor = 'var(--color-primary-700)';
-      activeBgColor = COLORS.background1;
-      activeColor = COLORS.font1;
-      disabledBorderColor = 'var(--color-grey-400)';
-      disabledBgColor = 'var(--color-grey-000)';
+      colors.borderColor = 'var(--color-grey-700)';
+      colors.backgroundColor = 'var(--color-grey-000)';
+      colors.color = 'var(--color-grey-900)';
+      colors.hoverBorderColor = 'var(--color-primary-500)';
+      colors.hoverBgColor = 'var(--color-grey-000)';
+      colors.hoverColor = 'var(--color-grey-900)';
+      colors.activeBorderColor = 'var(--color-primary-700)';
+      colors.activeBgColor = 'var(--color-grey-000)';
+      colors.activeColor = 'var(--color-grey-900)';
+      colors.disabledBorderColor = 'var(--color-grey-400)';
+      colors.disabledBgColor = 'var(--color-grey-000)';
+      break;
+
+    case 'nitrogen':
+      colors.borderColor = 'var(--color-grey-400)';
+      colors.backgroundColor = 'var(--color-grey-000)';
+      colors.color = 'var(--color-grey-900)';
+      colors.hoverBorderColor = 'var(--color-primary-500)';
+      colors.hoverBgColor = 'var(--color-grey-000)';
+      colors.hoverColor = 'var(--color-grey-900)';
+      colors.activeBorderColor = 'var(--color-primary-700)';
+      colors.activeBgColor = 'var(--color-grey-000)';
+      colors.activeColor = 'var(--color-grey-900)';
+      colors.disabledBorderColor = 'var(--color-grey-400)';
+      colors.disabledBgColor = 'var(--color-grey-000)';
       break;
 
     case 'helium':
-      borderColor = COLORS.primary1;
-      backgroundColor = COLORS.primary1;
-      color = COLORS.background1;
+      colors.borderColor = 'var(--color-primary-500)';
+      colors.backgroundColor = 'var(--color-primary-500)';
+      colors.color = 'var(--color-grey-000)';
       break;
 
     case 'lithium':
     case 'checked':
-      borderColor = 'var(--color-primary-500)';
-      backgroundColor = 'var(--color-primary-100)';
-      color = COLORS.font1;
-      hoverBorderColor = 'var(--color-primary-700)';
-      hoverBgColor = 'var(--color-primary-300)';
-      hoverColor = COLORS.font1;
-      activeBorderColor = 'var(--color-primary-900)';
-      activeBgColor = 'var(--color-primary-300)';
-      activeColor = COLORS.font1;
-      disabledBorderColor = 'var(--color-grey-600)';
-      disabledBgColor = 'var(--color-grey-300)';
+      colors.borderColor = 'var(--color-primary-500)';
+      colors.backgroundColor = 'var(--color-primary-100)';
+      colors.color = 'var(--color-grey-900)';
+      colors.hoverBorderColor = 'var(--color-primary-700)';
+      colors.hoverBgColor = 'var(--color-primary-300)';
+      colors.hoverColor = 'var(--color-grey-900)';
+      colors.activeBorderColor = 'var(--color-primary-900)';
+      colors.activeBgColor = 'var(--color-primary-300)';
+      colors.activeColor = 'var(--color-grey-900)';
+      colors.disabledBorderColor = 'var(--color-grey-600)';
+      colors.disabledBgColor = 'var(--color-grey-300)';
       break;
 
     case 'beryllium':
-      borderColor = COLORS.font1;
-      backgroundColor = COLORS.font1;
-      color = COLORS.background1;
+      colors.borderColor = 'var(--color-grey-900)';
+      colors.backgroundColor = 'var(--color-grey-900)';
+      colors.color = 'var(--color-grey-000)';
       break;
 
     case 'copper':
-      borderColor = 'var(--color-danger-300)';
-      backgroundColor = COLORS.background1;
-      color = 'var(--color-danger-500)';
-      hoverBorderColor = 'var(--color-danger-500)';
-      hoverBgColor = COLORS.background1;
-      hoverColor = 'var(--color-danger-500)';
-      activeBorderColor = 'var(--color-danger-700)';
-      activeBgColor = COLORS.background1;
-      activeColor = 'var(--color-danger-700)';
-      disabledBorderColor = 'var(--color-grey-400)';
-      disabledBgColor = 'var(--color-grey-000)';
+      colors.borderColor = 'var(--color-danger-300)';
+      colors.backgroundColor = 'var(--color-grey-000)';
+      colors.color = 'var(--color-danger-500)';
+      colors.hoverBorderColor = 'var(--color-danger-500)';
+      colors.hoverBgColor = 'var(--color-grey-000)';
+      colors.hoverColor = 'var(--color-danger-500)';
+      colors.activeBorderColor = 'var(--color-danger-700)';
+      colors.activeBgColor = 'var(--color-grey-000)';
+      colors.activeColor = 'var(--color-danger-700)';
+      colors.disabledBorderColor = 'var(--color-grey-400)';
+      colors.disabledBgColor = 'var(--color-grey-000)';
       break;
 
     case 'boron':
-      borderColor = COLORS.grey1;
-      backgroundColor = COLORS.font1;
-      color = COLORS.background1;
-      hoverBorderColor = 'currentColor';
-      hoverBgColor = COLORS.line3;
-      hoverColor = COLORS.background1;
-      activeBorderColor = 'currentColor';
-      activeBgColor = COLORS.line3;
-      activeColor = COLORS.font2;
+      colors.borderColor = 'var(--color-grey-700)';
+      colors.backgroundColor = 'var(--color-grey-900)';
+      colors.color = 'var(--color-grey-000)';
+      colors.hoverBorderColor = 'currentColor';
+      colors.hoverBgColor = 'var(--color-grey-800)';
+      colors.hoverColor = 'var(--color-grey-000)';
+      colors.activeBorderColor = 'currentColor';
+      colors.activeBgColor = 'var(--color-grey-800)';
+      colors.activeColor = 'var(--color-grey-500)';
       break;
 
     case 'scandium':
-      borderColor = COLORS.font1;
-      backgroundColor = COLORS.font1;
-      color = COLORS.font3;
-      hoverBorderColor = COLORS.line3;
-      hoverBgColor = COLORS.line3;
-      hoverColor = COLORS.background1;
-      activeBorderColor = COLORS.grey1;
-      activeBgColor = COLORS.grey1;
-      activeColor = COLORS.background1;
+      colors.borderColor = 'var(--color-grey-900)';
+      colors.backgroundColor = 'var(--color-grey-900)';
+      colors.color = 'var(--color-grey-600)';
+      colors.hoverBorderColor = 'var(--color-grey-800)';
+      colors.hoverBgColor = 'var(--color-grey-800)';
+      colors.hoverColor = 'var(--color-grey-000)';
+      colors.activeBorderColor = 'var(--color-grey-700)';
+      colors.activeBgColor = 'var(--color-grey-700)';
+      colors.activeColor = 'var(--color-grey-000)';
       break;
 
     case 'neon':
-      borderColor = COLORS.orange;
-      backgroundColor = COLORS.orange;
-      color = COLORS.background1;
-      hoverBorderColor = COLORS.orange2;
-      hoverBgColor = COLORS.orange2;
-      hoverColor = COLORS.background1;
-      activeBorderColor = COLORS.orange3;
-      activeBgColor = COLORS.orange3;
-      activeColor = COLORS.background1;
+      colors.borderColor = 'var(--color-warning-500)';
+      colors.backgroundColor = 'var(--color-warning-500)';
+      colors.color = 'var(--color-grey-000)';
+      colors.hoverBorderColor = 'var(--color-warning-700)';
+      colors.hoverBgColor = 'var(--color-warning-700)';
+      colors.hoverColor = 'var(--color-grey-000)';
+      colors.activeBorderColor = 'var(--color-warning-900)';
+      colors.activeBgColor = 'var(--color-warning-900)';
+      colors.activeColor = 'var(--color-grey-000)';
       break;
 
     case 'iron':
-      borderColor = 'var(--color-danger-500)';
-      backgroundColor = 'var(--color-danger-500)';
-      color = COLORS.background1;
-      hoverBorderColor = 'var(--color-danger-700)';
-      hoverBgColor = 'var(--color-danger-700)';
-      hoverColor = COLORS.background1;
-      activeBorderColor = 'var(--color-danger-900)';
-      activeBgColor = 'var(--color-danger-900)';
-      activeColor = COLORS.background1;
+      colors.borderColor = 'var(--color-danger-500)';
+      colors.backgroundColor = 'var(--color-danger-500)';
+      colors.color = 'var(--color-grey-000)';
+      colors.hoverBorderColor = 'var(--color-danger-700)';
+      colors.hoverBgColor = 'var(--color-danger-700)';
+      colors.hoverColor = 'var(--color-grey-000)';
+      colors.activeBorderColor = 'var(--color-danger-900)';
+      colors.activeBgColor = 'var(--color-danger-900)';
+      colors.activeColor = 'var(--color-grey-000)';
       break;
 
     case 'krypton':
-      borderColor = 'var(--color-success-500)';
-      backgroundColor = 'var(--color-success-500)';
-      color = COLORS.background1;
-      hoverBorderColor = 'var(--color-success-700)';
-      hoverBgColor = 'var(--color-success-700)';
-      hoverColor = COLORS.background1;
-      activeBorderColor = 'var(--color-success-900)';
-      activeBgColor = 'var(--color-success-900)';
-      activeColor = COLORS.background1;
+      colors.borderColor = 'var(--color-success-500)';
+      colors.backgroundColor = 'var(--color-success-500)';
+      colors.color = 'var(--color-grey-000)';
+      colors.hoverBorderColor = 'var(--color-success-700)';
+      colors.hoverBgColor = 'var(--color-success-700)';
+      colors.hoverColor = 'var(--color-grey-000)';
+      colors.activeBorderColor = 'var(--color-success-900)';
+      colors.activeBgColor = 'var(--color-success-900)';
+      colors.activeColor = 'var(--color-grey-000)';
       break;
   }
 
-  return css(["border:", " ", " ", ";background-color:", ";color:", ";svg:not(.k-ColorSvg){&,path{fill:currentColor;transition:fill var(--transition);}}transition:background-color var(--transition),color var(--transition),border-color var(--transition);", " &:disabled,&.k-Button--disabled{border:var(--border-width) solid ", ";background-color:", ";color:", ";}"], borderSize, borderStyle, borderColor, backgroundColor, color, modifier !== 'checked' && css(["&:hover{border-color:", ";background-color:", ";color:", ";}&:focus{border-color:", ";background-color:", ";color:", ";}&:active{border-color:", ";background-color:", ";color:", ";}"], hoverBorderColor, hoverBgColor, hoverColor, hoverBorderColor, backgroundColor, color, activeBorderColor, activeBgColor, activeColor), disabledBorderColor, disabledBgColor, disabledColor);
+  return css(["border:var(--border-width) solid ", ";background-color:", ";color:", ";svg:not(.k-ColorSvg){&,path{fill:currentColor;transition:fill var(--transition);}}transition:background-color var(--transition),color var(--transition),border-color var(--transition);", " &:disabled,&.k-Button--disabled{border:var(--border-width) solid ", ";background-color:", ";color:", ";}"], colors.borderColor, colors.backgroundColor, colors.color, modifier !== 'checked' && css(["&:hover{border-color:", ";background-color:", ";color:", ";}&:focus{border-color:", ";background-color:", ";color:", ";}&:active{border-color:", ";background-color:", ";color:", ";}"], colors.hoverBorderColor, colors.hoverBgColor, colors.hoverColor, colors.hoverBorderColor, colors.backgroundColor, colors.color, colors.activeBorderColor, colors.activeBgColor, colors.activeColor), colors.disabledBorderColor, colors.disabledBgColor, colors.disabledColor);
 };